setting up of Controller Article, Comment, Author with config and delete masterController
