+++ src/Model/Session.php
@@ -1,0 +1,21 @@
+<?php 
+
+class Session{
+
+    public function __construct()
+    {
+        session_start();
+        
+    }
+
+    public function setFlash($type, $message){
+        
+    }
+
+    public function showFlash(){
+
+    }
+
+
+
+}
