+++ templates/admin/comment-edit.html.twig
@@ -51,7 +51,7 @@
 						</div>
 						<br>
 						<!--Permet de récupérer l'id du commentaire -->
-						<input name="modComment" type="hidden" value="{{comment.id_pk_comment}}">
+						<input name="edit_comment" type="hidden" value="{{comment.id_pk_comment}}">
 
 						<div class="form-group">
 							<button type="submit" class="btn btn-primary">Modifier</button>
