+++ templates/admin/registration-valid.html.twig
@@ -36,8 +36,8 @@
                                 <strong>Role : </strong> {{author.role}}
                             </div>
                             <div class="card-footer">
-                                <a href="?page=registration-valid&amp;validAuthor={{author.id_pk_author}}" class="btn btn-success">Valider</a>
-								<a href="?page=registration-valid&amp;deleteAuthor={{author.id_pk_author}}" class="btn btn-danger">Supprimer</a>
+                                <a href="?page=registration-valid&amp;valid_author={{author.id_pk_author}}" class="btn btn-success">Valider</a>
+								<a href="?page=registration-valid&amp;delete_author={{author.id_pk_author}}" class="btn btn-danger">Supprimer</a>
                                
                             </div>
                         </div>
