+++ templates/home.html.twig
@@ -15,7 +15,12 @@
 					<div class="site-heading">
 						<h1>Blog
 						</h1>
-						<span class="subheading">La solution qu'il vous faut !</span>
+						<span class="subheading">La solution qu'il vous faut !</span><br><br>
+						{% if session.message %}
+										
+						<p class="bg-success p-3">{{session.message}}</p>
+						
+						{% endif %}
 					</div>
 				</div>
 			</div>
