+++ templates/member/login.html.twig
@@ -14,7 +14,8 @@
 					<div class="site-heading">
 						<h1>
 							Connexion
-						</h1>
+						</h1>			
+						
 					</div>
 				</div>
 			</div>
@@ -28,7 +29,9 @@
 		<div class="row">
 			<div class="col-md-6 mx-auto d-block bg-light p-5">
 				<form class = "form-group" method="POST">
-
+					{% if session.message %}								
+					<p class="bg-danger p-3">{{session.message}}</p>
+					{% endif %}					
 					<div class="form-group">
 						<label for="email">Email</label>
 						<input name="email" type="email" class="form-control" placeholder="Saisissez votre email..." required>
@@ -46,5 +49,7 @@
 			</div>
 		</div>
 	</div>
+
+	
 
 {% endblock %}
