+++ templates/member/registration.html.twig
@@ -48,7 +48,7 @@
 						<label for="confirmPassword">Confirmation mot de passe</label>
 						<input name="confirmPassword" type="password" class="form-control" placeholder="Confirmez le mot de passe.." required>
 					</div>
-					<input type="hidden" name="addAuthor" value ="1">
+					<input type="hidden" name="add_author" value ="1">
 					<button type="submit" class="btn btn-success">S'inscrire</button>
 					
 				</form>
