+++ templates/partials/header.html.twig
@@ -12,7 +12,7 @@
 		<div class="collapse navbar-collapse" id="navbarResponsive">
 			<ul class="navbar-nav ml-auto">
 				<li class="nav-item">
-					<a class="nav-link" href="?page=home">Accueil</a>
+					<a class="nav-link" href="?page=home">Accueil</a>				
 				</li>
 				{% if session.role == '' %}				
 				<li class="nav-item ">
