+++ templates/post-list.html.twig
@@ -29,6 +29,7 @@
 						<a href="?page=post-detail&amp;id_article={{author.id_pk_article}}&amp;id_author={{author.id_author}}">
 							<h2 class="post-title">
 								{{author.title}}
+								
 							</h2>
 							<h3 class="post-subtitle">
 								{{author.sentence}}
